# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/posts/popular": {
      "get": {
        "operationId": "PostController_listPopular",
        "summary": "List popular buddy posts",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/EventDetailPostDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts": {
      "get": {
        "operationId": "PostController_list",
        "summary": "List posts by activity or current user",
        "parameters": [
          {
            "name": "activityLegacyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "anchorPostId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated posts for an activity",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/EventPostsPageDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "PostController_create",
        "summary": "Create a buddy post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostMutationResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/posts/ai-search": {
      "post": {
        "operationId": "PostController_aiSearch",
        "summary": "Natural language buddy post search",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiSearchPostsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BuddyPostAiSearchResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/posts/{id}/recruit": {
      "patch": {
        "operationId": "PostController_updateRecruit",
        "summary": "Update post recruit status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostRecruitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostMutationResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/posts/{id}": {
      "patch": {
        "operationId": "PostController_update",
        "summary": "Update post content",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostMutationResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PostController_remove",
        "summary": "Delete owned post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostMutationResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/posts/{id}/comments": {
      "get": {
        "operationId": "PostController_listComments",
        "summary": "List comments for a post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostCommentsPageDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "PostController_addComment",
        "summary": "Add comment to post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostMutationResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/posts/{id}/comments/{commentId}": {
      "delete": {
        "operationId": "PostController_removeComment",
        "summary": "Delete owned comment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostMutationResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Sync API",
    "description": "Sync App REST API. Responses are wrapped by TransformInterceptor as { code, message, data }.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiEnvelopeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "description": "Response payload"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
...
      "EventDetailPostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "avatar": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "recruitStatus": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "location",
          "tags",
          "avatar"
        ]
      },
      "EventPostsPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDetailPostDto"
            }
          },
          "nextCursor": {
            "type": "string"
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "hasMore"
        ]
      },
      "BuddyPostAiSearchResultDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventDetailPostDto"
            }
          },
          "totalMatched": {
            "type": "number"
          },
          "totalScanned": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "totalMatched",
          "totalScanned"
        ]
      },
      "AiSearchPostsDto": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1
          },
          "activityLegacyId": {
            "type": "number"
          }
        },
        "required": [
          "query",
          "activityLegacyId"
        ]
      },
      "PostMutationResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "id"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1
          },
          "activityLegacyId": {
            "type": "number"
          },
          "eventTitle": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "departureCity": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "listedInFeed": {
            "type": "boolean",
            "description": "Default true. False = save post but hide from activity/popular feeds."
          },
          "recruitStatus": {
            "type": "string",
            "enum": [
              "open",
              "full"
            ]
          },
          "slotsTotal": {
            "type": "number",
            "minimum": 1
          },
          "slotsFilled": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "body"
        ]
      },
      "UpdatePostRecruitDto": {
        "type": "object",
        "properties": {
          "recruitStatus": {
            "type": "string",
            "enum": [
              "open",
              "full"
            ]
          },
          "slotsTotal": {
            "type": "number"
          },
          "slotsFilled": {
            "type": "number"
          }
        },
        "required": [
          "recruitStatus"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1
          },
          "location": {
            "type": "string"
          },
          "departureCity": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "recruitStatus": {
            "type": "string",
            "enum": [
              "open",
              "full"
            ]
          },
          "slotsTotal": {
            "type": "number",
            "minimum": 1
          },
          "slotsFilled": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "body"
        ]
      },
      "PostCommentItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "time": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "authorName",
          "avatar",
          "body",
          "time"
        ]
      },
      "PostCommentsPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostCommentItemDto"
            }
          },
          "hasMore": {
            "type": "boolean"
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items",
          "hasMore"
        ]
      },
      "CreatePostCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1
          },
          "parentCommentId": {
            "type": "string"
          }
        },
        "required": [
          "body"
        ]
      },
...
    }
  }
}